requirements.txt
capstone==5.0.1
keystone-engine==0.9.2
pefile==2023.2.7
